DOMINIC FEB11,2025 9:36:

+++ database/migrations/2025_01_30_024326_create_patents_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('patents', function (Blueprint $table) {
             $table->increments('PatentID');
-            $table->unsignedBigInteger('DiscID')->nullable;
+            $table->unsignedInteger('DiscID')->nullable;
             $table->foreign ('DiscID')
             ->references('DiscID')
             ->on('Disclosure');
