+++ database/migrations/2025_01_30_025549_create_trademarks_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('trademarks', function (Blueprint $table) {
             $table->increments('TrademarkID');
-            $table->unsignedBigInteger('DiscID')->nullable;
+            $table->unsignedInteger('DiscID')->nullable;
             $table->foreign ('DiscID')
             ->references('DiscID')
             ->on('Disclosure');
