+++ database/migrations/2025_01_30_031349_create_copyrights_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('copyrights', function (Blueprint $table) {
             $table->increments('Copyright ID');
-            $table->unsignedBigInteger('DiscID')->nullable;
+            $table->unsignedInteger('DiscID')->nullable;
             $table->foreign ('DiscID')
             ->references('DiscID')
             ->on('Disclosure');
